requirements.txt
Flask_WTF==0.15.1
Flask_Bcrypt==0.7.1
flask_paginate==0.8.1
Flask_Login==0.5.0
Flask==2.0.1
Flask_SQLAlchemy==2.5.1
WTForms==2.3.3
python-decouple==3.5
gunicorn==19.9.0